push admin

+++ src/main/java/org/example/cdio/dto/ProductInventoryView.java
@@ -1,12 +1,10 @@
 package org.example.cdio.dto;
 
 import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.Setter;
-
-import java.time.LocalDateTime;
-
 import lombok.Data;
+
+import java.math.BigDecimal;
+import java.time.LocalDateTime;
 
 @Data
 @AllArgsConstructor
@@ -15,11 +13,10 @@ public class ProductInventoryView {
     private Long productId;
     private String productName;
     private Integer weightGrams;
-    private Double wholesalePrice;
+    private BigDecimal wholesalePrice;
     private Integer quantity;
     private LocalDateTime createdAt;
     private Integer expiryDays;
     private LocalDateTime expiredAt;
     private Long daysLeft;
-
 }
